+++ tests/flat-bpc/convertToFlatBpcGraph.test.ts
@@ -39,7 +39,7 @@ test("convertToFlatBpcGraph generates correct nodes and edges", () => {
   expect(flat.nodes.length).toBe(5)
 
   // N1 has 2 pins → 1 edge, N2 single pin → 0 edges
-  expect(flat.undirectedEdges.length).toBe(1)
+  expect(flat.undirectedEdges.length).toBe(2)
   expect(flat.undirectedEdges[0]).toEqual(["B1-P1", "B2-P2"])
 
   const boxNode = flat.nodes.find((n) => n.id === "B1")!
